requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
Flask-Mail==0.9.1
werkzeug==2.3.7 